+++ x/evm/keeper/keeper.go
@@ -18,31 +18,43 @@ import (
 	"github.com/tharsis/ethermint/x/evm/types"
 )
 
-// Keeper wraps the CommitStateDB, allowing us to pass in SDK context while adhering
-// to the StateDB interface.
+// Keeper grants access to the EVM module state and implements the go-ethereum StateDB interface.
 type Keeper struct {
 	// Protobuf codec
 	cdc codec.BinaryCodec
 	// Store key required for the EVM Prefix KVStore. It is required by:
-	// - storing Account's Storage State
-	// - storing Account's Code
+	// - storing account's Storage State
+	// - storing account's Code
 	// - storing transaction Logs
-	// - storing block height -> bloom filter map. Needed for the Web3 API.
+	// - storing Bloom filters by block height. Needed for the Web3 API.
 	storeKey sdk.StoreKey
 
 	// key to access the transient store, which is reset on every block during Commit
 	transientKey sdk.StoreKey
 
-	paramSpace    paramtypes.Subspace
+	// module specific parameter space that can be configured through governance
+	paramSpace paramtypes.Subspace
+	// access to account state
 	accountKeeper types.AccountKeeper
-	bankKeeper    types.BankKeeper
+	// update balance and accounting operations with coins
+	bankKeeper types.BankKeeper
+	// access historical headers for EVM state transition execution
 	stakingKeeper types.StakingKeeper
 
-	ctx sdk.Context
+	// Manage the initial context and cache context stack for accessing the store,
+	// emit events and log info.
+	// It is kept as a field to make is accessible by the StateDb
+	// functions. Resets on every transaction/block.
+	ctxStack ContextStack
 
 	// chain ID number obtained from the context's chain id
 	eip155ChainID *big.Int
-	debug         bool
+
+	// Tracer used to collect execution traces from the EVM transaction execution
+	tracer string
+	// trace EVM state transition execution. This value is obtained from the `--trace` flag.
+	// For more info check https://geth.ethereum.org/docs/dapp/tracing
+	debug bool
 }
 
 // NewKeeper generates new evm module keeper
@@ -50,7 +62,7 @@ func NewKeeper(
 	cdc codec.BinaryCodec,
 	storeKey, transientKey sdk.StoreKey, paramSpace paramtypes.Subspace,
 	ak types.AccountKeeper, bankKeeper types.BankKeeper, sk types.StakingKeeper,
-	debug bool,
+	tracer string, debug bool,
 ) *Keeper {
 
 	// ensure evm module account is set
@@ -72,8 +84,24 @@ func NewKeeper(
 		stakingKeeper: sk,
 		storeKey:      storeKey,
 		transientKey:  transientKey,
+		tracer:        tracer,
 		debug:         debug,
 	}
+}
+
+// Ctx returns the current context from the context stack
+func (k Keeper) Ctx() sdk.Context {
+	return k.ctxStack.CurrentContext()
+}
+
+// CommitCachedContexts commit all the cache contexts created by `StateDB.Snapshot`.
+func (k *Keeper) CommitCachedContexts() {
+	k.ctxStack.Commit()
+}
+
+// CachedContextsEmpty returns true if there's no cache contexts.
+func (k *Keeper) CachedContextsEmpty() bool {
+	return k.ctxStack.IsEmpty()
 }
 
 // Logger returns a module-specific logger.
@@ -81,9 +109,9 @@ func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", types.ModuleName)
 }
 
-// WithContext sets an updated SDK context to the keeper
+// WithContext clears the context stack, and set the initial context.
 func (k *Keeper) WithContext(ctx sdk.Context) {
-	k.ctx = ctx
+	k.ctxStack.Reset(ctx)
 }
 
 // WithChainID sets the chain id to the local variable in the keeper
@@ -178,8 +206,8 @@ func (k Keeper) SetBlockBloom(ctx sdk.Context, height int64, bloom ethtypes.Bloo
 
 // GetBlockBloomTransient returns bloom bytes for the current block height
 func (k Keeper) GetBlockBloomTransient() *big.Int {
-	store := prefix.NewStore(k.ctx.TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
-	heightBz := sdk.Uint64ToBigEndian(uint64(k.ctx.BlockHeight()))
+	store := prefix.NewStore(k.Ctx().TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
+	heightBz := sdk.Uint64ToBigEndian(uint64(k.Ctx().BlockHeight()))
 	bz := store.Get(heightBz)
 	if len(bz) == 0 {
 		return big.NewInt(0)
@@ -191,8 +219,8 @@ func (k Keeper) GetBlockBloomTransient() *big.Int {
 // SetBlockBloomTransient sets the given bloom bytes to the transient store. This value is reset on
 // every block.
 func (k Keeper) SetBlockBloomTransient(bloom *big.Int) {
-	store := prefix.NewStore(k.ctx.TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
-	heightBz := sdk.Uint64ToBigEndian(uint64(k.ctx.BlockHeight()))
+	store := prefix.NewStore(k.Ctx().TransientStore(k.transientKey), types.KeyPrefixTransientBloom)
+	heightBz := sdk.Uint64ToBigEndian(uint64(k.Ctx().BlockHeight()))
 	store.Set(heightBz, bloom.Bytes())
 }
 
@@ -202,7 +230,7 @@ func (k Keeper) SetBlockBloomTransient(bloom *big.Int) {
 
 // GetTxHashTransient returns the hash of current processing transaction
 func (k Keeper) GetTxHashTransient() common.Hash {
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	bz := store.Get(types.KeyPrefixTransientTxHash)
 	if len(bz) == 0 {
 		return common.Hash{}
@@ -213,13 +241,13 @@ func (k Keeper) GetTxHashTransient() common.Hash {
 
 // SetTxHashTransient set the hash of processing transaction
 func (k Keeper) SetTxHashTransient(hash common.Hash) {
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	store.Set(types.KeyPrefixTransientTxHash, hash.Bytes())
 }
 
 // GetTxIndexTransient returns EVM transaction index on the current block.
 func (k Keeper) GetTxIndexTransient() uint64 {
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	bz := store.Get(types.KeyPrefixTransientTxIndex)
 	if len(bz) == 0 {
 		return 0
@@ -232,7 +260,7 @@ func (k Keeper) GetTxIndexTransient() uint64 {
 // value by one and then sets the new index back to the transient store.
 func (k Keeper) IncreaseTxIndexTransient() {
 	txIndex := k.GetTxIndexTransient()
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	store.Set(types.KeyPrefixTransientTxIndex, sdk.Uint64ToBigEndian(txIndex+1))
 }
 
@@ -266,7 +294,7 @@ func (k Keeper) GetAllTxLogs(ctx sdk.Context) []types.TransactionLogs {
 // GetLogs returns the current logs for a given transaction hash from the KVStore.
 // This function returns an empty, non-nil slice if no logs are found.
 func (k Keeper) GetTxLogs(txHash common.Hash) []*ethtypes.Log {
-	store := prefix.NewStore(k.ctx.KVStore(k.storeKey), types.KeyPrefixLogs)
+	store := prefix.NewStore(k.Ctx().KVStore(k.storeKey), types.KeyPrefixLogs)
 
 	bz := store.Get(txHash.Bytes())
 	if len(bz) == 0 {
@@ -281,7 +309,7 @@ func (k Keeper) GetTxLogs(txHash common.Hash) []*ethtypes.Log {
 
 // SetLogs sets the logs for a transaction in the KVStore.
 func (k Keeper) SetLogs(txHash common.Hash, logs []*ethtypes.Log) {
-	store := prefix.NewStore(k.ctx.KVStore(k.storeKey), types.KeyPrefixLogs)
+	store := prefix.NewStore(k.Ctx().KVStore(k.storeKey), types.KeyPrefixLogs)
 
 	txLogs := types.NewTransactionLogsFromEth(txHash, logs)
 	bz := k.cdc.MustMarshal(&txLogs)
@@ -297,7 +325,7 @@ func (k Keeper) DeleteTxLogs(ctx sdk.Context, txHash common.Hash) {
 
 // GetLogSizeTransient returns EVM log index on the current block.
 func (k Keeper) GetLogSizeTransient() uint64 {
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	bz := store.Get(types.KeyPrefixTransientLogSize)
 	if len(bz) == 0 {
 		return 0
@@ -310,7 +338,7 @@ func (k Keeper) GetLogSizeTransient() uint64 {
 // value by one and then sets the new index back to the transient store.
 func (k Keeper) IncreaseLogSizeTransient() {
 	logSize := k.GetLogSizeTransient()
-	store := k.ctx.TransientStore(k.transientKey)
+	store := k.Ctx().TransientStore(k.transientKey)
 	store.Set(types.KeyPrefixTransientLogSize, sdk.Uint64ToBigEndian(logSize+1))
 }
 
@@ -339,7 +367,7 @@ func (k Keeper) GetAccountStorage(ctx sdk.Context, address common.Address) (type
 // ----------------------------------------------------------------------------
 
 func (k Keeper) DeleteState(addr common.Address, key common.Hash) {
-	store := prefix.NewStore(k.ctx.KVStore(k.storeKey), types.AddressStoragePrefix(addr))
+	store := prefix.NewStore(k.Ctx().KVStore(k.storeKey), types.AddressStoragePrefix(addr))
 	key = types.KeyAddressStorage(addr, key)
 	store.Delete(key.Bytes())
 }
@@ -360,22 +388,22 @@ func (k Keeper) DeleteCode(addr common.Address) {
 		return
 	}
 
-	store := prefix.NewStore(k.ctx.KVStore(k.storeKey), types.KeyPrefixCode)
+	store := prefix.NewStore(k.Ctx().KVStore(k.storeKey), types.KeyPrefixCode)
 	store.Delete(hash.Bytes())
 }
 
 // ClearBalance subtracts the EVM all the balance denomination from the address
 // balance while also updating the total supply.
 func (k Keeper) ClearBalance(addr sdk.AccAddress) (prevBalance sdk.Coin, err error) {
-	params := k.GetParams(k.ctx)
+	params := k.GetParams(k.Ctx())
 
-	prevBalance = k.bankKeeper.GetBalance(k.ctx, addr, params.EvmDenom)
+	prevBalance = k.bankKeeper.GetBalance(k.Ctx(), addr, params.EvmDenom)
 	if prevBalance.IsPositive() {
-		if err := k.bankKeeper.SendCoinsFromAccountToModule(k.ctx, addr, types.ModuleName, sdk.Coins{prevBalance}); err != nil {
+		if err := k.bankKeeper.SendCoinsFromAccountToModule(k.Ctx(), addr, types.ModuleName, sdk.Coins{prevBalance}); err != nil {
 			return sdk.Coin{}, stacktrace.Propagate(err, "failed to transfer to module account")
 		}
 
-		if err := k.bankKeeper.BurnCoins(k.ctx, types.ModuleName, sdk.Coins{prevBalance}); err != nil {
+		if err := k.bankKeeper.BurnCoins(k.Ctx(), types.ModuleName, sdk.Coins{prevBalance}); err != nil {
 			return sdk.Coin{}, stacktrace.Propagate(err, "failed to burn coins from evm module account")
 		}
 	}